# server/matching/generated/src/main/java/org/openapitools/api/GenAiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ConversationStarterCollection;
import org.openapitools.model.UserCollection;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-07-19T16:53:54.581009086Z[Etc/UTC]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "GenAI", description = "Paths belonging to the GenAI microservice")
public interface GenAiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/v2/genai/conversation-starter : Request conversation starter
     * Request a series of conversation starter prompts from the GenAI microservice. Provide infomation about users on request.
     *
     * @param userCollection Request Conversation starter for these users (optional)
     * @return OK (status code 200)
     *         or The request was malformed or contained invalid parameters.  (status code 400)
     *         or Authentication failed due to missing or invalid OAuth2 token.  (status code 401)
     *         or The requested resource was not found.  (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getApiV2GenaiConversationStarter",
        summary = "Request conversation starter",
        description = "Request a series of conversation starter prompts from the GenAI microservice. Provide infomation about users on request.",
        tags = { "GenAI" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ConversationStarterCollection.class))
            }),
            @ApiResponse(responseCode = "400", description = "The request was malformed or contained invalid parameters. "),
            @ApiResponse(responseCode = "401", description = "Authentication failed due to missing or invalid OAuth2 token. "),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found. "),
            @ApiResponse(responseCode = "500", description = "Internal Server Error")
        },
        security = {
            @SecurityRequirement(name = "jwt-bearer")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v2/genai/conversation-starter",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ConversationStarterCollection> getApiV2GenaiConversationStarter(
        @Parameter(name = "UserCollection", description = "Request Conversation starter for these users") @Valid @RequestBody(required = false) @Nullable UserCollection userCollection
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"conversationsStarters\" : [ { \"prompt\" : \"prompt\" }, { \"prompt\" : \"prompt\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}