# server/user/generated-client/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: devoops@tyrean.mozmail.com
    name: Team DevOops
    url: https://github.com/AET-DevOps25/team-devoops/
  description: "This OpenAPI specification defines the endpoints, schemas, and security\
    \ mechanisms\nfor the Meet@Mensa User micro-service. "
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: ""
  title: MeetAtMensa
  version: 2.3.0
servers:
- description: ""
  url: api.meetatmensa.com
security:
- jwt-bearer: []
...
paths:
...
  /api/v2/genai/conversation-starter:
    get:
      description: Request a series of conversation starter prompts from the GenAI
        microservice. Provide infomation about users on request.
      operationId: get-api-v2-genai-conversation-starter
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCollection"
        description: Request Conversation starter for these users
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationStarterCollection"
          description: OK
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: Request conversation starter
      tags:
      - GenAI
      x-stoplight:
        id: 2gb78xli5e55w
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
...
  responses:
    UnauthorizedError:
      description: |
        Authentication failed due to missing or invalid OAuth2 token.
    NotFoundError:
      description: |
        The requested resource was not found.
    BadRequestError:
      description: |
        The request was malformed or contained invalid parameters.
  schemas:
    ConversationStarter:
      description: Object representing a conversation starter in the meet@mensa system
      example:
        prompt: prompt
      properties:
        prompt:
          type: string
          x-stoplight:
            id: lmwqzkbspxydt
      required:
      - prompt
      title: ConversationStarter
      x-stoplight:
        id: 5grlrzedrmo9k
    ConversationStarterCollection:
      description: Object representing a collection of conversation starters in the
        Meet@Mensa system.
      example:
        conversationsStarters:
        - prompt: prompt
        - prompt: prompt
      properties:
        conversationsStarters:
          items:
            $ref: "#/components/schemas/ConversationStarter"
          type: array
          x-stoplight:
            id: b82xoshonnop6
      required:
      - conversationsStarters
      title: ConversationStarterCollection
      x-stoplight:
        id: p1vhmcpea8fmp
...
    interest:
      description: String representing an interest a user has
      title: interest
      type: string
      x-stoplight:
        id: 2gt8tcnnuxxjw
...
    User:
      description: Object representing a student user in the Meet@Mensa system.
      example:
        birthday: 2000-01-23
        firstname: Max
        degreeStart: 2024
        gender: other
        degree: msc_informatics
        bio: bio
        interests:
        - null
        - null
        userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        lastname: Mustermann
      properties:
        userID:
          description: The unique ID of a single student in the Meet@Mensa system.
          format: uuid
          title: userID
          type: string
          x-stoplight:
            id: qlawgs207p4xu
        email:
          description: Users's e-mail
          format: email
          type: string
          x-stoplight:
            id: e7l4vk5tcpzog
        firstname:
          description: User's given name
          example: Max
          type: string
          x-stoplight:
            id: d32i2omc3k6nn
        lastname:
          description: User's surname
          example: Mustermann
          type: string
          x-stoplight:
            id: 5jc0p1zr8vy4p
        birthday:
          description: User's date of birth
          format: date
          type: string
          x-stoplight:
            id: nm996scnjz7dk
        gender:
          default: other
          description: User's gender
          type: string
          x-stoplight:
            id: itop4gfsf1ch7
        degree:
          description: User's degree program
          example: msc_informatics
          type: string
          x-stoplight:
            id: u7kkx1ewi0g3o
        degreeStart:
          description: What year User started their degree
          example: 2024
          maximum: 2100
          minimum: 2000
          type: integer
          x-stoplight:
            id: nd14ato08ps5k
        interests:
          description: Array of a User's interests
          items:
            $ref: "#/components/schemas/interest"
          type: array
          x-stoplight:
            id: y9ktoyz2v1to4
        bio:
          description: Short introduction text written by the user
          type: string
          x-stoplight:
            id: pyp2f4rynj24v
      required:
      - bio
      - birthday
      - degree
      - degreeStart
      - email
      - firstname
      - gender
      - interests
      - lastname
      - userID
      title: User
      x-stoplight:
        id: k21d7pds0kt8x
    UserCollection:
      description: Object representing a collection of student user in the Meet@Mensa
        system.
      example:
        users:
        - birthday: 2000-01-23
          firstname: Max
          degreeStart: 2024
          gender: other
          degree: msc_informatics
          bio: bio
          interests:
          - null
          - null
          userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          lastname: Mustermann
        - birthday: 2000-01-23
          firstname: Max
          degreeStart: 2024
          gender: other
          degree: msc_informatics
          bio: bio
          interests:
          - null
          - null
          userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          lastname: Mustermann
      properties:
        users:
          items:
            $ref: "#/components/schemas/User"
          type: array
          x-stoplight:
            id: l8a1hngdfmgzr
      required:
      - users
      title: UserCollection
      x-stoplight:
        id: 8qy8pit8jdvq3
...
  securitySchemes:
    jwt-bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
...